Added constructor for class node and method to insert multiple objects of the same value.
Added a constructor for class node to make the AVL class completely independent from the structure of node. The only mandatory fields are data (the value of the object), count (the number of objects with this value), balance and maxdepth (necessary to maintain structure of AVL). Other fields can be added and manipulated striclty within the node class. This makes the class great for tackling problems in which we need to maintain other types of information in the tree. For instance, field "subtree" can be omitted as it is only used in the "kth_element" method.